Inject update checker to telegram-notify

+++ telegram-notify/install.sh
@@ -13,6 +13,19 @@ if [ -f "$SPATH/telegram-notify" ]; then
 	cp -v $SPATH/telegram-notify /usr/local/bin/telegram-notify && \
 	chmod +x /usr/local/bin/telegram-notify
 
+	if [ -f "$SPATH/telegram-notify-updatecheck" ]; then
+		echo "Installing update checker..."
+
+		MERGE=`tail -n +3 "$SPATH/telegram-notify-updatecheck"`
+		MATCH=`awk '/#   Notification/{ print NR; exit }' /usr/local/bin/telegram-notify`
+
+		if [ "$MATCH" != "" ]; then
+			echo "$({ head -n $(($MATCH-2)) /usr/local/bin/telegram-notify; echo "$MERGE"; tail -n +$(($MATCH-2)) /usr/local/bin/telegram-notify; })" > /usr/local/bin/telegram-notify
+		else
+			echo "Update checker installation failed - cannot find anchor point to insert the code"
+		fi
+	fi
+
 	if [ ! -f "/etc/telegram-notify.conf" ] && [ -f "$SPATH/telegram-notify.conf" ]; then
 		cp -v $SPATH/telegram-notify.conf /etc/telegram-notify.conf
 	fi
@@ -20,6 +33,8 @@ if [ -f "$SPATH/telegram-notify" ]; then
 	if [ -f "$SPATH/telegram-notify@.service" ]; then
 		cp -v $SPATH/telegram-notify@.service /etc/systemd/system/telegram-notify@.service
 	fi
+
+	command -v dos2unix >/dev/null 2>&1 && dos2unix /usr/local/bin/telegram-notify
 else
 	exit 1
 fi
